requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
requests
google-auth
Pillow
PyJWT

